Fix faceit request logging
Add response.status to the log entry. This gives us easy field to look
for when finding abnormal behaviour, e.g. error codes 4XX, 5XX.

Use response.text() instead of response.json() because the content
might not always be JSON. If there is an error and the response is not
valid JSON we still want to be able to log the response content and
plain text is the only reliable way to do that.