Fix one part of the broken build
When running `source nvm.sh` the script will try to find a suitable
`.nvmrc` file indicating the nodejs version to use. If the desired
version is not already available (it isn't) and needs to be installed
the script will exit with exit code 3 (error) instead of 0 (success).
Why? No one knows.

Since we used the `errexit` option (`set -e`) to exit on any non-zero
exit code to avoid all kinds of chaos when errors happen it also kills
the build.

With `|| true` we force the exit code to be 0 and therefore ignore
errors.