Use v6 API explicitly and fix a bug caused by obnoxious changes by Discord
Discord has made a stupid decision and always point non-versioned API
URLs to the newest API version. This means that they not only break
applications using the API but potentially cause totally unpredictable
changes in application behaviour.

For example our message archiver no longer works and we only found out
by accident. We check that the channels "type" attribute is "text" to
find out the channels that need archiving but because the API no longer
uses "text" but 0 instead to mark a channel as a guild text channel our
archiver has not archived anything since 2017-07-19.

https://discordapp.com/developers/docs/change-log